Move render mode conversion to TGFX helpers.
